Add the skeleton of the Python project for input
This part of this project will receive the signals from IR Receiver
from the serial sent by the AVR Atmega and I'll interpret that signals
as special commands (whatever I want) as a Python program to control
my personal computer.

It isn't that nice? Finally just passing the next episode of animes
and series. :D